Lift licence composition out of licence.scenario.js into an entity helper
Adds tests/support/entities/, a new home for the valid/minimum data an entity needs to exist in its entirety, starting with licence.entity.js. Every scenario that was building on top of licence.scenario.js as a nested dependency now calls the entity directly instead, so composing a licence no longer chains through another scenario file. licence.scenario.js stays as a thin wrapper so "just a licence" remains pickable in cli:seed.
