+++ tests/support/entities/README.md
@@ -1,0 +1,19 @@
+# Entities
+
+An entity file (`*.entity.js`) builds one thing completely — the valid, minimum data an entity needs to exist, with all the child records the API requires to accept it. It answers "what does a valid licence look like?", not "what does *this test* need a licence to look like?".
+
+This is different from `tests/support/data/*.data.js` and `tests/support/scenarios/*.scenario.js`:
+
+- A **data file** builds a single row for a single table (e.g. `licence.data.js` builds just the `licences` row).
+- An **entity file** composes several data files into the smallest valid whole (e.g. `licence.entity.js` composes a company and address as the licence holder, the licence itself, its permit licence, licence document, licence document header, licence document role, licence version, and a licence version purpose/point — because a licence isn't valid/queryable without them).
+- A **scenario file** composes data files, entity files, and/or other scenario files into whatever a specific spec needs to seed, and is what `cli:seed.cli.js` lists for manual seeding.
+
+## Why this exists
+
+Scenario files used to reach a "valid licence" by importing another scenario file (e.g. `licence-with-charge-version.scenario.js` importing `licence.scenario.js`, which `registered-licence.scenario.js` or a water-company variant might then build on top of again). That works, but it means the definition of "a valid licence" lives inside a scenario file — one entry in a flat, cli:seed-pickable list — that other scenarios then have to import as if it were just another building block, several levels deep.
+
+An entity file gives every scenario a single, direct place to get a valid instance of that entity, with no scenario-to-scenario chaining. A scenario file can still exist for the entity on its own (e.g. `licence.scenario.js` is a thin wrapper around `licence.entity.js`, kept so "just a licence" remains pickable in `cli:seed`), but anything composing *on top of* that entity should import the entity file directly, not the scenario file.
+
+## Adding a new entity
+
+Only pull an entity out of a scenario file once more than one scenario needs "a valid X" as a starting point (as happened with `licence`). A data file whose only caller is a single scenario doesn't need this — see `.agents/skills/scenarios/SKILL.md` for when a data file should be split, and `.agents/skills/standards/SKILL.md` for naming conventions.
