+++ tests/support/scenarios/licence-with-all-return-log-statuses.scenario.js
@@ -3,7 +3,7 @@ import returnRequirementData from '../data/return-requirement.data.js'
 import returnRequirementPointData from '../data/return-requirement-point.data.js'
 import returnRequirementPurposeData from '../data/return-requirement-purpose.data.js'
 import returnVersionData from '../data/return-version.data.js'
-import licenceScenario from './licence.scenario.js'
+import licenceEntity from '../entities/licence.entity.js'
 import { relativeToToday } from '../helpers/date.helpers.js'
 
 export const title = 'All return log statuses'
@@ -13,7 +13,7 @@ export default function (calculatedDates) {
   const currentPeriod = _currentPeriod(calculatedDates)
   const previousPeriod = _previousPeriod(currentPeriod)
 
-  const licence = licenceScenario()
+  const licence = licenceEntity()
   const returnVersion = returnVersionData(licence.licence)
 
   // In the service return logs will cover the whole period of their matching return version. To ensure our test data is
