+++ tests/support/scenarios/licence-with-two-purposes.scenario.js
@@ -1,7 +1,7 @@
 import licenceVersionPurposeData from '../data/licence-version-purpose.data.js'
 import licenceVersionPurposePointData from '../data/licence-version-purpose-point.data.js'
 import pointData from '../data/point.data.js'
-import licenceScenario from './licence.scenario.js'
+import licenceEntity from '../entities/licence.entity.js'
 import { regionCode } from '../default-values.js'
 
 export const title = 'Licence with two purposes'
@@ -9,7 +9,7 @@ export const description =
   'A licence with two points and two licence version purposes, and no existing return requirements'
 
 export default function () {
-  const licence = licenceScenario()
+  const licence = licenceEntity()
 
   const secondPoint = pointData()
