+++ tests/support/scenarios/licence-with-workflow.scenario.js
@@ -1,11 +1,11 @@
 import workflowData from '../data/workflow.data.js'
-import licenceScenario from './licence.scenario.js'
+import licenceEntity from '../entities/licence.entity.js'
 
 export const title = 'Licence in workflow'
 export const description = 'A licence, licence holder (company), and a workflow entry'
 
 export default function () {
-  const licence = licenceScenario()
+  const licence = licenceEntity()
   const workflow = workflowData(licence.licence)
 
   return {
