+++ tests/support/scenarios/registered-licence.scenario.js
@@ -1,12 +1,12 @@
 import primaryUserData from '../data/primary-user.data.js'
-import licenceScenario from './licence.scenario.js'
+import licenceEntity from '../entities/licence.entity.js'
 import { externalUserEmail } from '../default-values.js'
 
 export const title = 'Registered licence'
 export const description = 'A licence that has been registered (primary user), licence holder and a company'
 
 export default function () {
-  const licence = licenceScenario()
+  const licence = licenceEntity()
 
   const primaryUser = primaryUserData(externalUserEmail, licence.company)
